fix due to voteValue
